+++ index.es.js
@@ -48,23 +48,35 @@ export default function css (options = {}) {
       includePaths = includePaths.filter((v, i, a) => a.indexOf(v) === i)
       try {
         const sass = options.sass || loadSassLibrary()
-        const css = sass.renderSync(Object.assign({
+
+        const render = sass.renderSync(Object.assign({
           data: prefix + scss,
+          outFile: dest,
           includePaths
-        }, options)).css.toString()
+        }, options))
+
+        const css = render.css.toString()
+        const map = render.map ? render.map.toString() : null
+
         // Possibly process CSS (e.g. by PostCSS)
         if (typeof options.processor === 'function') {
-          const processor = options.processor(css, styles)
+          const processor = options.processor(css, map, styles)
 
           // PostCSS support
           if (typeof processor.process === 'function') {
-            return Promise.resolve(processor.process(css, { from: undefined }))
-              .then(result => result.css)
+            return Promise.resolve(processor.process(css, {
+              from: undefined,
+              to: dest,
+              map: map ? { prev: map, inline: false } : null
+            }))
           }
 
-          return processor
+          return Promise.resolve(processor).then(result => {
+            if (typeof result === 'string') return { css: result }
+            else return result
+          })
         }
-        return css
+        return { css, map }
       } catch (e) {
         if (options.failOnError) {
           throw e
@@ -109,7 +121,7 @@ export default function css (options = {}) {
         const files = Array.isArray(options.watch) ? options.watch : [options.watch]
         files.forEach(file => this.addWatchFile(file))
       }
-
+      
       if (options.insert === true) {
         // When the 'insert' is enabled, the stylesheet will be inserted into <head/> tag.
         return Promise.resolve(compileToCSS(code)).then((css) => ({
@@ -119,8 +131,8 @@ export default function css (options = {}) {
         }))
       } else if (options.output === false) {
         // When output is disabled, the stylesheet is exported as a string
-        return Promise.resolve(compileToCSS(code)).then(css => ({
-          code: 'export default ' + JSON.stringify(css),
+        return Promise.resolve(compileToCSS(code)).then(compiled => ({
+          code: 'export default ' + JSON.stringify(compiled.css),
           map: { mappings: '' }
         }))
       }
@@ -142,47 +154,60 @@ export default function css (options = {}) {
         scss += styles[id] || ''
       }
 
-      const css = compileToCSS(scss)
+      if (typeof dest !== 'string') {
+        // Guess destination filename
+        dest = opts.dest || opts.file || 'bundle.js'
+        if (dest.endsWith('.js')) {
+          dest = dest.slice(0, -3)
+        }
+        dest = dest + '.css'
+      }
+
+      const compiled = compileToCSS(scss)
 
       // Resolve if processor returned a Promise
-      Promise.resolve(css).then(css => {
+      Promise.resolve(compiled).then(compiled => {
+        if (typeof compiled !== 'object' || typeof compiled.css !== 'string') {
+          return
+        }
+
         // Emit styles through callback
         if (typeof options.output === 'function') {
-          options.output(css, styles)
+          options.output(compiled.css, styles)
           return
         }
 
-        if (typeof css !== 'string') {
+        // Don't create unwanted empty stylesheets
+        if (!compiled.css.length) {
           return
-        }
-
-        if (typeof dest !== 'string') {
-          // Don't create unwanted empty stylesheets
-          if (!css.length) {
-            return
-          }
-
-          // Guess destination filename
-          dest = opts.dest || opts.file || 'bundle.js'
-          if (dest.endsWith('.js')) {
-            dest = dest.slice(0, -3)
-          }
-          dest = dest + '.css'
         }
 
         // Ensure that dest parent folders exist (create the missing ones)
         ensureParentDirsSync(dirname(dest))
 
         // Emit styles to file
-        writeFile(dest, css, (err) => {
+        writeFile(dest, compiled.css, (err) => {
           if (opts.verbose !== false) {
             if (err) {
               console.error(red(err))
-            } else if (css) {
-              console.log(green(dest), getSize(css.length))
+            } else if (compiled.css) {
+              console.log(green(dest), getSize(compiled.css.length))
             }
           }
         })
+
+        if (options.sourceMap && compiled.map) {
+          let sourcemap = compiled.map
+          if (typeof compiled.map.toString === 'function') {
+            sourcemap = compiled.map.toString()
+          }
+
+          writeFile(dest + '.map', sourcemap, (err) => {
+            if (opts.verbose !== false && err) {
+              console.error(red(err))
+            }
+          })
+        }
       })
     }
   }