+++ test/sourcemap/rollup.config.js
@@ -1,6 +1,4 @@
 import scss from '../../index.es.js'
-import postcss from 'postcss'
-import autoprefixer from 'autoprefixer'
 
 export default {
   input: '../input.js',
@@ -10,7 +8,7 @@ export default {
   },
   plugins: [
     scss({
-      processor: () => postcss([autoprefixer({ overrideBrowserslist: 'Edge 18' })])
+      sourceMap: true
     })
   ]
 }